chore: extract WipeDevice modal form UserContextModal, its not needed to have it global
