+++ packages/suite/src/components/suite/modals/index.tsx
@@ -22,7 +22,6 @@ export { QrScannerModal } from './ReduxModal/UserContextModal/QrScannerModal';
 export { BackgroundGalleryModal } from './ReduxModal/UserContextModal/BackgroundGalleryModal';
 export { TxDetailModal } from './ReduxModal/UserContextModal/TxDetailModal/TxDetailModal';
 export { ApplicationLogModal } from './ReduxModal/UserContextModal/ApplicationLogModal';
-export { WipeDeviceModal } from './ReduxModal/UserContextModal/WipeDeviceModal';
 export { MetadataProviderModal } from './ReduxModal/UserContextModal/MetadataProviderModal/MetadataProviderModal';
 export { AdvancedCoinSettingsModal } from './ReduxModal/UserContextModal/AdvancedCoinSettingsModal/AdvancedCoinSettingsModal';
 export { AddTokenModal } from './ReduxModal/UserContextModal/AddTokenModal';
