+++ packages/suite/src/views/settings/SettingsDevice/WipeDeviceModal.tsx
